Add New Script

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -6,7 +6,7 @@
  *
  * Return: returns last value of a number
  */
-int print_last_digit(long int n)
+int print_last_digit(int n)
 {
 	int last_digit;
